feat: serve existing COG overviews (Overview strategy) (#88)
## What

Teaches the read path to serve the overview IFDs real COGs already
contain — the planner's third strategy (#37), served today by the
tiler's own selection. The `RasterSource` port grows a `ReadLevel`
selector (`FullRes | Overview { factor }`), the COG adapter reads the
chosen reduced-resolution IFD with real recorded provenance, and the
tiler picks the overview at decimating zooms, emitting
`Strategy::Overview { level }` in the Trace. No pyramid generation, no
raster writer. Closes #38.

## Why

- **Coordinate-space contract**: requests stay in **full-resolution
pixel coordinates** at every level; the adapter covers them onto the
overview grid (floor start / ceil end against the exact per-axis ratio,
then clips) and returns `WindowData::grid` — the overview grid actually
read (real IFD dimensions, geotransform scaled by the exact ratio,
rasterio-identical). The warp now takes its grid from `WindowData`
itself, so callers never do overview math and an overview window warps
correctly by construction.
- **Level naming**: decimation factors (2 = half res), exactly as
`describe`'s `overview_levels` reports them — self-describing, no
positional-index ambiguity. `Strategy::Overview.level` widened `u8 ->
u32` (JSON wire shape unchanged; pinned tests untouched).
- **Selection rule**: coarsest factor ≤ desired downsampling ratio × 1.2
— GDAL's oversampling threshold (`GDALBandGetBestOverviewLevel2`),
calibrated empirically: at z11 the desired ratio is ~1.97 and
rio-tiler's default path demonstrably serves the ×2 overview
(byte-identical to an explicit `overview_level=0` open), so a 1.0
threshold would diverge from the oracle. z12 (~0.99) stays full-res
Live. The tiler's selection is a documented seam #37's planner will
subsume; the Trace reports `Overview` only when every band that read
pixels used the same level.

## Validation

- **Adapter truth (byte-exact)**: `window_truth.py` now also emits
`overview_truth.json` from **explicit** rasterio overview opens
(`overview_level=0` — the same stored bytes async-tiff decodes): 25
cases across the 5 fixtures pinning the rounding contract, the overview
grid (dims + transform `60/-60`, shared origin), SHA-256 of the
little-endian pixels, nodata stats. All pass byte-exactly, plus
local/in-memory agreement, `OverviewNotFound` (factor 4 → available
`[2]`), empty out-of-bounds clips, and a bytes-reduction assertion.
Existing 25-case full-res truth regenerated **byte-identical** and
untouched-green.
- **Render goldens (default `DiffPolicy`)**: new z11 overview-path
goldens (`b04-ov-…`, `fmask-ov-…`) rendered by the oracle from the same
overview IFD via `compose --overview-level 0 --exact-grid`. Measured
through `render_tile` (auto-selected overview): **b04 max |diff| 1,
0.0000% bad; fmask byte-exact (max |diff| 0)**. One documented
deviation: the pure two-stage rio-tiler default-path golden fails the
default policy by a hair (max |diff| 8, 0.5188% bad vs 0.5% — its
out-shape nearest snap, the same read-path artifact the existing z11
kernel goldens dodge), so the golden pins GDAL's single-stage warp of
the identical overview bytes, per the `--exact-grid` precedent;
`--overview-level` selection itself is verified byte-identical to
rio-tiler's own default overview path. The existing `--no-overviews
--exact-grid` z11 kernel goldens are untouched.
- **Trace / bytes reduction**: z11 b04 reports `Overview { level: 2 }`
with `bytes_read` **69,591 vs 272,973** for the identical render forced
full-res (**3.9× reduction**, asserted loosely at 2×); provenance
non-empty and inside the overview IFD's ranges. SSE overview envelope
shape pinned alongside live/cache_hit; the x-ray overlay already renders
`overview (level N)`.
- **Gates**: full `just check` green (unmasked exit 0) after rebasing
over #87; `just render-goldens` reproduces every committed golden
byte-identically; `just oracle-verify` and `just fixtures-verify` pass;
`just e2e` green — z12 tiles remain full-res Live, ingest-to-pixel 789
ms.

## Checklist

- [x] `just check` passes locally
- [x] Commits are DCO signed-off (`git commit -s`)
- [x] Docs/ADRs updated if behavior or a decision changed (module docs
carry the contracts; no ADR-level decision changed)

---------

+++ crates/swath-core/src/trace.rs
@@ -41,10 +41,14 @@ pub enum Strategy {
     },
     /// Pixels came from a pre-computed overview pyramid level.
     Overview {
-        /// Overview level read. The numbering convention (which level is
-        /// coarsest) is source-defined; the trace records the level as the
-        /// adapter reported it.
-        level: u8,
+        /// Overview level read, recorded as the **decimation factor** of
+        /// the overview grid relative to full resolution (2 = half
+        /// resolution, 4 = quarter, …) — exactly the naming
+        /// `RasterInfo::overview_levels` and `ReadLevel::Overview` use,
+        /// so the x-ray shows the same number the port speaks. (Widened
+        /// from `u8` to `u32` when overview reads landed in #38; the JSON
+        /// wire shape — a plain number — is unchanged.)
+        level: u32,
     },
     /// Pixels came from full-resolution source reads.
     Live,
